Initial implementation of spsolve, reach, and dfs.
NOTE Test fails. There are multiple issues with the raw pointers used in
the original C code vs. using a more object-oriented C++ solution. The
code segfaults when using `xi.size() == 2*N`, which the book describes
as the correct size. It runs with `xi.size() == N`, but gives an
incorrect result.