Rewrite dfs to use explicit stacks. Returns output in reverse order.
TODO do we want to reverse the actual output? Or use a modified version
of the `head` pointer to build the output list? We really only need the
output from `reach` to be in the "correct" order, not `dfs`, since we
make consecutive calls to `dfs`. It may be worth the O(N) cost to just
flip the order on output from `reach`.