Tests of lsolve_perm and usolve_perm pass.
We haven't yet figured out how to merge the two into tri_solve_perm.
There is no known way to determine whether the matrix is upper or lower
triangular, because we need to know if the permutation vectors are
reversed or not.

There should be a way to get the correct permuted indices and iterate
through in the correct order (aka reversed column order for U) without
needing to know if the matrix is upper or lower triangular.