Complete Exercise 4.12: rechol given symbolic pattern of L.
Not sure if we actually did this correctly? The only lines I changed
were removing the allocation of L, and removing any assignments to L.i_.
There may be further optimizations we can do given that we know the
locations of the non-zeros in L, not just the column counts.