Use LAPACK Householder method in cs::house. Outputs not consistent.
No other changes made in the cs::sqr, cs::vcount, or cs::qr code. I am
not sure why this method does not produce correct results, since it is
just using a vector normal to the Davis Householder vector. There seems
to be only a sign difference in the last column of Q and the diagonal of
R. See python/scripts/qr_example.py.