Change csint to std::int32_t. Numpy casting now works properly.
There was a weird bug in some tests where a CSCMatrix was getting
corrupted data, but silently. The numpy types of
a scipy.csc_array.indices and indptr on my system are np.int32, so there
was a mismatch when using csint = std::int64_t.

csc_array.data is np.float64 which corresponds to double, and is fine.