Refactor cspy to use matplotlib Rectangles instead of scatter.
`scatter` is plotted in screen coordinates, not data coordinates, so zooming
in keeps the points the same size (which is too small). `imshow` is
plotted in data coordinates, but requires a dense array, which can take
up a massive amount of memory for a large input matrix. Use
a `PatchCollection` to plot Rectangle markers that fill the integer
space appropriately in data coordinates, so zooming works.